{"schema":"decodex/commit/1","summary":"Move window content policy to owner","authority":"manual"}
