+++ packages/rsnap-overlay/src/overlay/window_content_policy.rs
@@ -1,0 +1,1 @@
+pub(in crate::overlay) const CAPTURE_WINDOW_CONTENT_PROTECTION_ENABLED: bool = false;
